Fix FlattenDependencies not skipping packages
If a package is referred to be a specifier such as "pc:foo" instead
of its actual name, we would never skip that as already resolved
as we tracked it by package name. In some rare cases this can cause
an infinite loop. Fix by simply deferring the "already resolved"
check until we've resolved the name, and use that.